% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_applyButtonInput.R
\name{mod_applyButtonInput}
\alias{mod_applyButtonInput}
\alias{mod_applyButton}
\title{mod_applyButtonInput and mod_applyButton}
\usage{
mod_applyButtonInput(id, lang, texts_thes)

mod_applyButton(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny module to create and populate the buttons inputs
}
