% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_buttonsInput.R
\name{mod_buttonsInput}
\alias{mod_buttonsInput}
\alias{mod_buttons}
\title{mod_buttonsInput and mod_buttons}
\usage{
mod_buttonsInput(id, nfidb)

mod_buttons(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny module to create and populate the buttons inputs
}
