% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_dataInput.R
\name{mod_data}
\alias{mod_data}
\title{mod_data server function}
\usage{
mod_data(input, output, session, nfidb, var_thes, texts_thes)
}
\arguments{
\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{nfidb}{pool object to access the nfi db}

\item{var_thes}{variables thesaurus df}
}
\description{
mod_data server function
}
