% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_infoUI.R
\name{mod_info}
\alias{mod_info}
\title{mod_info server function}
\usage{
mod_info(input, output, session, map_inputs, data_inputs, nfidb, var_thes,
  texts_thes)
}
\arguments{
\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{map_inputs}{map module reactives, including returned data}

\item{data_inputs}{data inputs module reactives}

\item{nfidb}{pool object with the db connection}

\item{var_thes}{variables thesaurus df}
}
\description{
mod_info server function
}
