% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_glossaryUI.R, R/mod_mapUI.R
\name{mod_glossary}
\alias{mod_glossary}
\alias{mod_mapUI}
\alias{mod_map}
\title{mod_map server function}
\usage{
mod_glossary(input, output, session, nfidb, var_thes, numerical_thes,
  texts_thes, lang, data_inputs)

mod_mapUI(id, nfidb)

mod_map(input, output, session, data_inputs, nfidb, var_thes, texts_thes,
  numerical_thes, lang, cache_list)
}
\arguments{
\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{nfidb}{pool with database connection object}

\item{var_thes}{variables thesaurus df}

\item{numerical_thes}{numerical variables df}

\item{texts_thes}{app translations df}

\item{lang}{reactive with the decalred lang}

\item{data_inputs}{reactive with the reactive data and the data inputs}

\item{id}{shiny id}

\item{cache_list}{inmemory cache}

\item{nfidb}{pool object to access the database}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{nfidb}{pool with database connection object}

\item{var_thes}{variables thesaurus df}
}
\description{
Shiny module to generate the map
}
