% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_returnedDataOutput.R
\name{mod_returnedData}
\alias{mod_returnedData}
\title{mod_returnedData server function}
\usage{
mod_returnedData(input, output, session, data_inputs, map_inputs = NULL,
  nfidb)
}
\arguments{
\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{data_inputs}{reactives from dataInput module}

\item{map_inputs}{map input with custom_polygon sf object, if any}

\item{nfidb}{pool object to access the nfi db}
}
