% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_returnedDataOutput.R
\name{mod_returnedDataOutput}
\alias{mod_returnedDataOutput}
\title{mod_returnedDataOutput and mod_returnedData}
\usage{
mod_returnedDataOutput(id)
}
\arguments{
\item{id}{}
}
\description{
Shiny module to get the data as tbl_sql
}
