% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_saveMapInput.R
\name{mod_saveMapInput}
\alias{mod_saveMapInput}
\alias{mod_saveMap}
\title{mod_saveMapInput and mod_saveMap}
\usage{
mod_saveMapInput(id, nfidb, lang)

mod_saveMap(input, output, session, map_inputs)
}
\arguments{
\item{id}{shiny id}

\item{nfidb}{pool object to access the database}

\item{lang}{lang value}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{map_inputs}{Reactive with the map reactives, including map_data}
}
\description{
A shiny module to create and populate the buttons inputs

mod_saveMap server function
}
