% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tableOutput.R
\name{mod_tableOutput}
\alias{mod_tableOutput}
\alias{mod_table}
\title{mod_tableOutput and mod_table}
\usage{
mod_tableOutput(id)

mod_table(input, output, session, data_inputs, map_inputs, nfidb, var_thes)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{data_inputs}{reactive with the reactive data and the data inputs}

\item{map_inputs}{reactive with the mod_map inputs, included main_data}

\item{nfidb}{db pool}

\item{var_thes}{variables thesaurus df}
}
\description{
A shiny module to generate the base IFN plots table

mod_table server function
}
