% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfi_app.R
\name{nfi_app}
\alias{nfi_app}
\title{function to launch the nfi app}
\usage{
nfi_app(user = "guest", password = "guest", host = NULL, port = NULL,
  dbname = NULL)
}
\description{
function to launch the nfi app
}
