% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_filtersUI.R
\name{mod_filtersUI}
\alias{mod_filtersUI}
\alias{mod_filters}
\title{mod_filtersUI and mod_filters}
\usage{
mod_filtersUI(id, nfidb)

mod_filters(input, output, session, nfidb, mod_data, apply_data, apply_viz)
}
\arguments{
\item{id}{shiny id}

\item{nfidb}{pool object to access the nfi db}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{mod_data}{reactives from the mod_dataInput module, to know about which scenario
we are}

\item{apply_data}{reactive from the applyButton module located in the data}

\item{apply_viz}{reactive from the applyButton module located in the viz}

\item{nfidb}{pool object to access the nfi db}
}
\description{
A shiny module to generate and process the filters

mod_filters server function
}
