% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_filtersUI.R
\name{mod_filtersUI}
\alias{mod_filtersUI}
\alias{mod_filters}
\title{mod_filtersUI and mod_filters}
\usage{
mod_filtersUI(id, nfidb, lang)

mod_filters(input, output, session, nfidb, data_inputs, lang)
}
\arguments{
\item{id}{shiny id}

\item{nfidb}{pool object to access the nfi db}

\item{lang}{lang value}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{data_inputs}{reactives from the data_inputsInput module, to know about which scenario
we are}

\item{nfidb}{pool object to access the nfi db}

\item{lang}{lang value}
}
\description{
A shiny module to generate and process the filters

mod_filters server function
}
