% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_mapUI.R
\name{mod_mapUI}
\alias{mod_mapUI}
\alias{mod_map}
\title{mod_mapUI and mod_map}
\usage{
mod_mapUI(id, nfidb)

mod_map(input, output, session, data_inputs, nfidb, lang)
}
\arguments{
\item{id}{shiny id}

\item{nfidb}{pool object to access the database}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{data_inputs}{reactive with the reactive data and the data inputs}

\item{lang}{lang value}

\item{nfidb}{pool with database connection object}
}
\description{
Shiny module to generate the map

mod_map server function
}
