% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_vizInput.R
\name{mod_viz}
\alias{mod_viz}
\title{mod_viz server function}
\usage{
mod_viz(input, output, session, data_inputs, nfidb, var_thes)
}
\arguments{
\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{data_inputs}{reactive with the reactive data and the data inputs}

\item{nfidb}{pool object to access nfi db}

\item{var_thes}{variables thesaurus df}
}
\description{
mod_viz server function
}
